+++ src/serial/hal_1.rs
@@ -12,8 +12,16 @@ impl<USART, WORD> ErrorType for super::Tx<USART, WORD> {
     type Error = super::Error;
 }
 
+impl<USART, RX, WORD> ErrorType for super::URx<USART, RX, WORD> {
+    type Error = super::Error;
+}
+
+impl<USART, TX, WORD> ErrorType for super::UTx<USART, TX, WORD> {
+    type Error = super::Error;
+}
+
 mod nb {
-    use super::super::{Error, Instance, Rx, Serial, Tx};
+    use super::super::{Error, Instance, Rx, Serial, Tx, URx, UTx};
     use embedded_hal_one::serial::{
         nb::{Read, Write},
         ErrorType,
@@ -35,12 +43,24 @@ mod nb {
         }
     }
 
+    impl<USART: Instance, RX> Read<u8> for URx<USART, RX, u8> {
+        fn read(&mut self) -> nb::Result<u8, Self::Error> {
+            self.read()
+        }
+    }
+
     /// Reads 9-bit words from the UART/USART
     ///
     /// If the UART/USART was configured with `WordLength::DataBits9`, the returned value will contain
     /// 9 received data bits and all other bits set to zero. Otherwise, the returned value will contain
     /// 8 received data bits and all other bits set to zero.
     impl<USART: Instance> Read<u16> for Rx<USART, u16> {
+        fn read(&mut self) -> nb::Result<u16, Self::Error> {
+            self.read()
+        }
+    }
+
+    impl<USART: Instance, RX> Read<u16> for URx<USART, RX, u16> {
         fn read(&mut self) -> nb::Result<u16, Self::Error> {
             self.read()
         }
@@ -70,6 +90,16 @@ mod nb {
         }
     }
 
+    impl<USART: Instance, TX> Write<u8> for UTx<USART, TX, u8> {
+        fn write(&mut self, word: u8) -> nb::Result<(), Self::Error> {
+            self.write(word)
+        }
+
+        fn flush(&mut self) -> nb::Result<(), Self::Error> {
+            self.flush()
+        }
+    }
+
     /// Writes 9-bit words to the UART/USART
     ///
     /// If the UART/USART was configured with `WordLength::DataBits9`, the 9 least significant bits will
@@ -87,7 +117,7 @@ mod nb {
 }
 
 mod blocking {
-    use super::super::{Instance, Serial, Tx};
+    use super::super::{Instance, Serial, Tx, UTx};
     use super::ErrorType;
     use embedded_hal_one::serial::blocking::Write;
 
@@ -115,7 +145,27 @@ mod blocking {
         }
     }
 
+    impl<USART: Instance, TX> Write<u8> for UTx<USART, TX, u8> {
+        fn write(&mut self, bytes: &[u8]) -> Result<(), Self::Error> {
+            self.bwrite_all(bytes)
+        }
+
+        fn flush(&mut self) -> Result<(), Self::Error> {
+            self.bflush()
+        }
+    }
+
     impl<USART: Instance> Write<u16> for Tx<USART, u16> {
+        fn write(&mut self, slice: &[u16]) -> Result<(), Self::Error> {
+            self.bwrite_all(slice)
+        }
+
+        fn flush(&mut self) -> Result<(), Self::Error> {
+            self.bflush()
+        }
+    }
+
+    impl<USART: Instance, TX> Write<u16> for UTx<USART, TX, u16> {
         fn write(&mut self, slice: &[u16]) -> Result<(), Self::Error> {
             self.bwrite_all(slice)
         }